2087: Clear open file perferences in non-blocking way
