+++ app/src/main/java/com/amaze/filemanager/application/AppConfig.java
@@ -73,7 +73,6 @@ public class AppConfig extends GlideApplication {
   private UtilsHandler utilsHandler;
 
   private WeakReference<Context> mainActivityContext;
-  private WeakReference<Context> themedActivityContext;
   private static ScreenUtils screenUtils;
 
   private static AppConfig instance;
@@ -208,10 +207,6 @@ public class AppConfig extends GlideApplication {
     screenUtils = new ScreenUtils(activity);
   }
 
-  public void setThemedActivityContext(@NonNull Activity activityContext) {
-    themedActivityContext = new WeakReference<>(activityContext);
-  }
-
   public ScreenUtils getScreenUtils() {
     return screenUtils;
   }
@@ -219,11 +214,6 @@ public class AppConfig extends GlideApplication {
   @Nullable
   public Context getMainActivityContext() {
     return mainActivityContext.get();
-  }
-
-  @Nullable
-  public Context getThemedActivityContext() {
-    return themedActivityContext.get();
   }
 
   public ExplorerDatabase getExplorerDatabase() {
